requirements.txt
asttokens==2.4.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.13.1
fonttools==4.55.0
fsspec==2024.2.0
h5py==3.12.1
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.3
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
nir==1.0.4
nirtorch==1.0
numpy==1.26.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==10.2.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
pyzmq==26.2.0
six==1.16.0
snntorch==0.9.1
stack-data==0.6.3
sympy==1.13.1
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
wcwidth==0.2.13
